making setting index level and adding integ test flag random
