+++ server/src/test/java/org/opensearch/search/DefaultSearchContextTests.java
@@ -213,8 +213,7 @@ public class DefaultSearchContextTests extends OpenSearchTestCase {
                 false,
                 Version.CURRENT,
                 false,
-                executor,
-                false
+                executor
             );
             contextWithoutScroll.from(300);
             contextWithoutScroll.close();
@@ -256,8 +255,7 @@ public class DefaultSearchContextTests extends OpenSearchTestCase {
                 false,
                 Version.CURRENT,
                 false,
-                executor,
-                false
+                executor
             );
             context1.from(300);
             exception = expectThrows(IllegalArgumentException.class, () -> context1.preProcess(false));
@@ -327,8 +325,7 @@ public class DefaultSearchContextTests extends OpenSearchTestCase {
                 false,
                 Version.CURRENT,
                 false,
-                executor,
-                false
+                executor
             );
 
             SliceBuilder sliceBuilder = mock(SliceBuilder.class);
@@ -367,8 +364,7 @@ public class DefaultSearchContextTests extends OpenSearchTestCase {
                 false,
                 Version.CURRENT,
                 false,
-                executor,
-                false
+                executor
             );
             ParsedQuery parsedQuery = ParsedQuery.parsedMatchAllQuery();
             context3.sliceBuilder(null).parsedQuery(parsedQuery).preProcess(false);
@@ -403,8 +399,7 @@ public class DefaultSearchContextTests extends OpenSearchTestCase {
                 false,
                 Version.CURRENT,
                 false,
-                executor,
-                false
+                executor
             );
             context4.sliceBuilder(new SliceBuilder(1, 2)).parsedQuery(parsedQuery).preProcess(false);
             Query query1 = context4.query();
@@ -434,8 +429,7 @@ public class DefaultSearchContextTests extends OpenSearchTestCase {
                 false,
                 Version.CURRENT,
                 false,
-                executor,
-                false
+                executor
             );
             int numSlicesForPit = maxSlicesPerPit + randomIntBetween(1, 100);
             when(sliceBuilder.getMax()).thenReturn(numSlicesForPit);
@@ -532,8 +526,7 @@ public class DefaultSearchContextTests extends OpenSearchTestCase {
                 false,
                 Version.CURRENT,
                 false,
-                executor,
-                false
+                executor
             );
             assertThat(context.searcher().hasCancellations(), is(false));
             context.searcher().addQueryCancellation(() -> {});
